+++ pkg/dds/CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.9.1
 - [DAP] `runInTerminal` requests are now sent after first responding to the `launchRequest`.
 - [DAP] Skipped tests are now marked with `!` instead of `✓` in `Output` events.
+- Fixed a bug that was preventing clients from receiving `IsolateReload` events
+  (see https://dartbug.com/49491).
 
 # 2.9.0
 - Updated DDS protocol to version 1.6.